+++ src/safe/auth.rs
