+++ src/safe/mod.rs
@@ -5,7 +5,10 @@ use sn_api::{fetch::XorUrlBase, Safe};
 use std::{str::FromStr, sync::Arc, time::Duration};
 use tokio::sync::RwLock;
 
-use crate::{auth, files, keys};
+mod auth;
+mod files;
+mod keys;
+mod nrs;
 
 #[js_function(2)]
 fn constructor(ctx: CallContext) -> Result<JsUndefined> {
@@ -44,6 +47,7 @@ pub fn define_class(env: &Env) -> Result<JsFunction> {
                 .with_method(keys::create_preload_test_coins),
             Property::new(&env, "keys_balance_from_sk")?.with_method(keys::balance_from_sk),
             Property::new(&env, "files_container_create")?.with_method(files::container_create),
+            Property::new(&env, "nrs_map_container_create")?.with_method(nrs::map_container_create),
         ],
     )
 }
